Create README.md
I recently participated in BCG's Data Science job simulation on the Forage platform, and it was incredibly useful to understand what it might be like to participate on a data science and analytics team at BCG.

I worked on a project to create a customer churn analysis simulation using Python. This project built my data science skills in a real-world context, as well as my presentation skills through creating an executive summary of my findings for the team. 

Doing this program confirmed that I really enjoy working on data science projects and I'm excited to apply these skills on a data science team at a company like BCG.